third_party/blink/web_tests/accessibility/aria-hidden-with-elements.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body id="body">

<div id="container" aria-label="container">

  <table border=1 summary="test" aria-hidden="true">
    <tr><td>cell</td><td>cell</td><td>cell</td></tr>
    <tr><td>cell</td><td>cell</td><td>cell</td></tr>
  </table>

  <div role="grid" aria-hidden="true">
    <div role="row">
      <div role="gridcell">cell</div>
    </div>
  </div>

  <ul aria-hidden="true">
  <li>test
  <li>test
  </ul>

  <div aria-label="wrapper">
    <div role="listbox" aria-hidden="true">
      <div role="option">option</div>
      <div role="option">option</div>
    </div>
  </div>

  <h1>heading</h1>

</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that aria-hidden works as expected on elements that are subclasses of AccessibilityRenderObject");

   if (window.accessibilityController) {
      var container = accessibilityController.accessibleElementById("container");

      // first child is an AXGroup that has 0 unignored children
      var child = container.childAtIndex(0);
      shouldBe("child.childrenCount", "6")
      for(let i = 0; i < child.childrenCount; ++i) {
        shouldBeTrue("child.childAtIndex(" + i +").isIgnored");
      }

      let i = 1;
      for(; i < container.childrenCount; ++i) {
        child = container.childAtIndex(i);
       if (!child.isIgnored)
       break;
      }
      shouldBe("child.role", "'AXRole: AXGenericContainer'");
      child = container.childAtIndex(++i);
      shouldBe("child.role", "'AXRole: AXHeading'");
    }

</script>

</body>
</html>
