third_party/blink/web_tests/accessibility/aria-hidden-with-elements-expected.txt

cell	cell	cell
cell	cell	cell
cell
test
test
option
option
heading

This tests that aria-hidden works as expected on elements that are subclasses of AccessibilityRenderObject

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS child.childrenCount is 6
PASS child.childAtIndex(0).isIgnored is true
PASS child.childAtIndex(1).isIgnored is true
PASS child.childAtIndex(2).isIgnored is true
PASS child.childAtIndex(3).isIgnored is true
PASS child.childAtIndex(4).isIgnored is true
PASS child.childAtIndex(5).isIgnored is true
PASS child.role is 'AXRole: AXGenericContainer'
PASS child.role is 'AXRole: AXHeading'
PASS successfullyParsed is true

TEST COMPLETE
